+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # Weblog of stuff
+
+- Added Twitter card and Telegram link previews to all vault pages (2026-02-10)